Create LocationMaster
I'm sorry for the Spaghetti Code

Here's how it works:
Locations are individual Locations - the contain String ArrayLists holding the list of adjacent locations, the list of npcs, the list of objects, and a name
LocationConstructor manually creates each Location, and currently has some placeholders.
LocationMaster is instantiated and immediately calls each LocationConstructor function, creating the ganme locations, and storing these locations in an arraylist. You call parseLocation, giving it a String. parseLocation then goes through the list of Locations, checking the provided String to each Location's name, and returns the Location with a matching name. If there is no matching Location, it returns a placeHolder value and prints that there wasn't a matching Location.

I don't fully understand GUI, so I've provided what I think should be enough to make a few adjustments to.
To move, a player first calls printNeighbor() in Location, to print out the names of the bordering Locations to the current Location. Then you call setLocation() in PlayerState, giving it the String name of the desired Location. setLocation checks if the String name it has been provided is in the neighbor arraylist of the currentLocation, and parses that name to get the Location object. If the two Locations border each other, the PlayerState's Location _location is changed to the desired Location. Otherwise, it prints an error. 

If I have a full list of Locations and their member data, I can maybe get this to run smoother with the 2d array, but for now this is my minimum deliverable on movement